GCC barfs this to the .s file directly:
    .objc_class_name_CrashTestPlugin=0
    .globl .objc_class_name_CrashTestPlugin

without creating a tree node or anything. We, on the other hand, are creating a
tree node and outputting:

        .globl .objc_class_name_CrashTestPlugin
        .zerofill __DATA, __common, .objc_class_name_CrashTestPlugin, 4, 2

This causes a linker crash because it's in the wrong section:

        (undefined [lazy bound]) external [no dead strip] .objc_class_name_NSObject
00000000 (__TEXT,__text) non-external _-[CrashTestPlugin crash]
000000d8 (__DATA,__common) external [no dead strip] .objc_class_name_CrashTestPlugin

The .objc_class_name_CrashTestPlugin symbol is bad.  It should look like:

00000000 (absolute) external .objc_class_name_CrashTestPlugin

Because it's barfed out directly to the file, we do the same with a file asm
directive.


git-svn-id: http://llvm.org/svn/llvm-project/llvm-gcc-4.2/trunk@46281 91177308-0d34-0410-b5e6-96231b3b80d8