Add SVR4 ABI support to llvm-gcc.
Create constant in rs6000_gimplify_va_arg() with a proper size.
If aggregates get decomposed, propagate the attributes of the aggregate to the scalars it is decomposed into.
Add custom ABI class for the SVR4 ABI. Integer arguments which need to be put on the stack are marked with the inreg attribute.


git-svn-id: http://llvm.org/svn/llvm-project/llvm-gcc-4.2/trunk@74785 91177308-0d34-0410-b5e6-96231b3b80d8