A clean-up needs to be marked as a clean-up. It could cause problems in the libunwind library otherwise. Take this program for an example:
int main() {
    try {
        throw new std::exception();
    } catch (std::exception *e) {
        throw e;
    }
}

At the re-throw, the libunwind expects the state to be in a clean-up
state. However, because we generate catch-alls in our code, the personality
returns the wrong state the libunwind, which then barfs.

Marking that re-throw as a clean-up puts libunwind into the correct state.

There is other code here, though. If we mark clean-ups as clean-ups, then it
could break when inlining through an 'invoke' instruction. You will get a
situation like this:

bb:
  %ehptr = eh.exception()
  %sel = eh.selector(%ehptr, @per, 0);

...

bb2:
  invoke _Unwind_Resume_or_Rethrow(%ehptr) %normal unwind to %lpad

lpad:
  ...

The unwinder will see the %sel call as a clean-up and, if it doesn't have a
catch further up the call stack, it will skip running it. But there *is* another
catch up the stack -- the catch for the %lpad. However, we can't see that. This
is fixed in code-gen, where we detect this situation, and convert the "clean-up"
selector call into a "catch-all" selector call. This gives us the correct
semantics.


git-svn-id: http://llvm.org/svn/llvm-project/llvm-gcc-4.2/trunk@99670 91177308-0d34-0410-b5e6-96231b3b80d8