Fix semantics of function-local asm "register" variables. Handling of these was allowing the value of the variable to be clobbered because the RAs allocated the register to something else.  8305081.
Both gcc and llvm-gcc refer to "register variables", but that
is not really what they are.  The real requirements are much looser:
"Defining such a register variable does not reserve the register; it
remains available for other uses in places where flow control determines
the variable's value is not live.

This option does not guarantee that GCC will generate code that has this 
variable in the register you specify at all times. You may not code an 
explicit reference to this register in the assembler instruction template 
part of an asm statement and assume it will always refer to this variable. 
However, using the variable as an asm operand guarantees that the specified 
register is used for the operand."

So, we can just allocate the variable normally, and use magical asm
pixie dust to get its value into the right register around asm's (in
practice, it's done after all loads, in the hope that this will help
optimizations keep the value in the appropriate register all the time when
there are enough registers to do that - works in small examples anyway).
The RA need not be bothered at all.  Clang doesn't currently support this 
feature at all but it should be able to imitate what llvm-gcc generates easily.

The existing code is retained so handling of global asm register variables
doesn't crash; they don't work any better than before, though.



git-svn-id: http://llvm.org/svn/llvm-project/llvm-gcc-4.2/trunk@112674 91177308-0d34-0410-b5e6-96231b3b80d8